+++ amzconnect-supporting-infra/lexbot_module/data.tf
@@ -1,7 +1,1 @@
 data "aws_caller_identity" "current" {}
-
-data "archive_file" "lexbot" {
-  type        = "zip"
-  source_dir  = "../lexbot"
-  output_path = "./ACME_lexbot.zip"
-}