+++ amzconnect-supporting-infra/lexbot_module/locals.tf
@@ -1,0 +1,4 @@
+locals {
+  lex_bot   = var.lexbot_complete_name
+  languages = var.lexbot_languages
+}